Rename Vultr _facts -> _info
