+++ changelogs/fragments/xxxxx-vultr-facts-info.yaml
@@ -1,0 +1,8 @@
+minor_changes:
+- The ``vultr_dns_domain_facts`` module has been deprecated. Use ``vultr_dns_domain_info`` instead.
+- The ``vultr_firewall_group_facts`` module has been deprecated. Use ``vultr_firewall_group_info`` instead.
+- The ``vultr_network_facts`` module has been deprecated. Use ``vultr_network_info`` instead.
+- The ``vultr_plan_facts`` module has been deprecated. Use ``vultr_plan_info`` instead.
+- The ``vultr_region_facts`` module has been deprecated. Use ``vultr_region_info`` instead.
+- The ``vultr_server_facts`` module has been deprecated. Use ``vultr_server_info`` instead.
+- The ``vultr_startup_script_facts`` module has been deprecated. Use ``vultr_startup_script_info`` instead.
