--- test/integration/targets/vultr_firewall_group_facts/tasks/main.yml
@@ -1,31 +1,0 @@
-# Copyright (c) 2018, Yanis Guenane <yanis+ansible@guenane.org>
-# GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
----
-- name: test gather vultr firewall group facts - empty resources
-  vultr_firewall_group_facts:
-
-- name: Create the firewall group
-  vultr_firewall_group:
-    name: '{{ firewall_group_name }}'
-
-- name: test gather vultr firewall group facts in check mode
-  vultr_firewall_group_facts:
-  check_mode: yes
-
-- name: verify test gather vultr firewall group facts in check mode
-  assert:
-    that:
-    - ansible_facts.vultr_firewall_group_facts|selectattr('description','equalto','{{ firewall_group_name }}') | list | count == 1
-
-- name: test gather vultr firewall group facts
-  vultr_firewall_group_facts:
-
-- name: verify test gather vultr firewall group facts
-  assert:
-    that:
-    - ansible_facts.vultr_firewall_group_facts|selectattr('description','equalto','{{ firewall_group_name }}') | list | count == 1
-
-- name: Delete the firewall group
-  vultr_firewall_group:
-    name: '{{ firewall_group_name }}'
-    state: absent
