+++ test/integration/targets/vultr_firewall_group_info/tasks/main.yml
@@ -1,0 +1,33 @@
+# Copyright (c) 2018, Yanis Guenane <yanis+ansible@guenane.org>
+# GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
+---
+- name: test gather vultr firewall group info - empty resources
+  vultr_firewall_group_info:
+
+- name: Create the firewall group
+  vultr_firewall_group:
+    name: '{{ firewall_group_name }}'
+
+- name: test gather vultr firewall group info in check mode
+  vultr_firewall_group_info:
+  check_mode: yes
+  register: result
+
+- name: verify test gather vultr firewall group info in check mode
+  assert:
+    that:
+    - result.vultr_firewall_group_info|selectattr('description','equalto','{{ firewall_group_name }}') | list | count == 1
+
+- name: test gather vultr firewall group info
+  vultr_firewall_group_info:
+  register: result
+
+- name: verify test gather vultr firewall group info
+  assert:
+    that:
+    - result.vultr_firewall_group_info|selectattr('description','equalto','{{ firewall_group_name }}') | list | count == 1
+
+- name: Delete the firewall group
+  vultr_firewall_group:
+    name: '{{ firewall_group_name }}'
+    state: absent
