--- test/integration/targets/vultr_network_facts/tasks/main.yml
@@ -1,33 +1,0 @@
-# Copyright (c) 2018, Yanis Guenane <yanis+ansible@guenane.org>
-# GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
----
-- name: test gather vultr network facts - empty resources
-  vultr_network_facts:
-
-- name: Create the network
-  vultr_network:
-    name: '{{ vultr_network_name }}'
-    cidr: '{{ vultr_network_cidr }}'
-    region: '{{ vultr_network_region }}'
-
-- name: test gather vultr network facts in check mode
-  vultr_network_facts:
-  check_mode: yes
-
-- name: verify test gather vultr network facts in check mode
-  assert:
-    that:
-    - ansible_facts.vultr_network_facts|selectattr('name','equalto','{{ vultr_network_name }}') | list | count == 1
-
-- name: test gather vultr network facts
-  vultr_network_facts:
-
-- name: verify test gather vultr network facts
-  assert:
-    that:
-    - ansible_facts.vultr_network_facts|selectattr('name','equalto','{{ vultr_network_name }}') | list | count == 1
-
-- name: Delete the script
-  vultr_network:
-    name: '{{ vultr_network_name }}'
-    state: absent
