+++ test/integration/targets/vultr_network_info/defaults/main.yml
