--- test/integration/targets/vultr_plan_facts/tasks/main.yml
@@ -1,19 +1,0 @@
-# Copyright (c) 2018, Yanis Guenane <yanis+ansible@guenane.org>
-# GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
----
-- name: test gather vultr plan facts in check mode
-  vultr_plan_facts:
-  check_mode: yes
-
-- name: verify test gather vultr plan facts in check mode
-  assert:
-    that:
-    - ansible_facts.vultr_plan_facts|selectattr('name','equalto','16384 MB RAM,2x110 GB SSD,20.00 TB BW') | list | count == 1
-
-- name: test gather vultr plan fact
-  vultr_plan_facts:
-
-- name: verify test gather vultr plan facts
-  assert:
-    that:
-    - ansible_facts.vultr_plan_facts|selectattr('name','equalto','16384 MB RAM,2x110 GB SSD,20.00 TB BW') | list | count == 1
