+++ test/integration/targets/vultr_plan_info/tasks/main.yml
@@ -1,0 +1,21 @@
+# Copyright (c) 2018, Yanis Guenane <yanis+ansible@guenane.org>
+# GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
+---
+- name: test gather vultr plan info in check mode
+  vultr_plan_info:
+  check_mode: yes
+  register: result
+
+- name: verify test gather vultr plan info in check mode
+  assert:
+    that:
+    - result.vultr_plan_info|selectattr('name','equalto','16384 MB RAM,2x110 GB SSD,20.00 TB BW') | list | count == 1
+
+- name: test gather vultr plan info
+  vultr_plan_info:
+  register: result
+
+- name: verify test gather vultr plan info
+  assert:
+    that:
+    - result.vultr_plan_info|selectattr('name','equalto','16384 MB RAM,2x110 GB SSD,20.00 TB BW') | list | count == 1
