--- test/integration/targets/vultr_region_facts/tasks/main.yml
@@ -1,19 +1,0 @@
-# Copyright (c) 2018, Yanis Guenane <yanis+ansible@guenane.org>
-# GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
----
-- name: test gather vultr region facts in check mode
-  vultr_region_facts:
-  check_mode: yes
-
-- name: verify test gather vultr region facts in check mode
-  assert:
-    that:
-    - ansible_facts.vultr_region_facts|selectattr('name','equalto','Atlanta') | list | count == 1
-
-- name: test gather vultr region facts
-  vultr_region_facts:
-
-- name: verify test gather vultr region facts
-  assert:
-    that:
-    - ansible_facts.vultr_region_facts|selectattr('name','equalto','Atlanta') | list | count == 1
