+++ test/integration/targets/vultr_region_info/aliases
