+++ test/integration/targets/vultr_server_info/defaults/main.yml
