+++ test/integration/targets/vultr_startup_script_info/aliases
