+++ test/integration/targets/vultr_startup_script_info/defaults/main.yml
